backend/requirements.txt
# ============================================================================
# requirements.txt — Week 6
# ============================================================================
# NEW this week:
#   reportlab   — pure-Python PDF generation (works on Windows)
#   python-docx — pure-Python Word doc generation (works on Windows)
# Both are pure Python — no system dependencies to worry about.
# ============================================================================

flask>=3.0
flask-cors>=4.0
flask-login>=0.6
flask-sqlalchemy>=3.1
bcrypt>=4.0
python-dotenv>=1.0
werkzeug>=3.0

pandas>=2.0
numpy>=1.24
scipy>=1.11
scikit-learn>=1.3
statsmodels>=0.14
pingouin>=0.5

faker>=25.0
openpyxl>=3.1

anthropic>=0.39

# NEW in Week 6 — report export
reportlab>=4.0
python-docx>=1.1
